refactor MQTT message handling
